Avoid potential number overflows. Found by VS code analyser
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@6393 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CB3DMeshWriter.cpp
@@ -98,7 +98,7 @@ bool CB3DMeshWriter::writeMesh(io::IWriteFile* file, IMesh* const mesh, s32 flag
     u32 numTexture = texs.size();
     for (u32 i = 0; i < numTexture; i++)
 	{
-        file->write(texs[i].TextureName.c_str(), texs[i].TextureName.size() + 1);
+        file->write(texs[i].TextureName.c_str(), (size_t)texs[i].TextureName.size() + 1);
         file->write(&texs[i].Flags, 7*4);
     }
